requirements.txt
streamlit==1.28.1
pymongo==4.6.0
pandas==2.1.4
plotly==5.17.0
Pillow==10.1.0
bcrypt==4.1.2
python-dateutil==2.8.2